+++ Frameworks/OgreKit/Headers/OgreTextViewUndoer.h
@@ -20,7 +20,7 @@
     NSMutableArray      *_attributedStringArray;
     unsigned            _count, _tail;
 }
-- (id)initWithCapacity:(unsigned)aCapacity;
+- (id)initWithCapacity:(NSUInteger)aCapacity;
 - (void)addRange:(NSRange)aRange attributedString:(NSAttributedString*)anAttributedString;
 - (void)undoTextView:(id)aTarget jumpToSelection:(BOOL)jumpToSelection invocationTarget:(id)myself;
 @end
